Refactor metadata storage implementations,
Add row_to_metadata_scan_item function to metadata_storage/sqlx module.
Make both postgres, sqlite impl codes to use sqlx::row_to_metadata_scan_item in scan_metadata impl.
